package/python-empy: new package
A powerful and robust templating system for Python.

EmPy is a system for embedding Python expressions and statements in
template text; it takes an EmPy source file, processes it, and produces
output. This is accomplished via expansions, which are special signals
to the EmPy system and are set off by a special prefix (by default the
at sign, @). EmPy can expand arbitrary Python expressions and statements
in this way, as well as a variety of special forms. Textual data not
explicitly delimited in this way is sent unaffected to the output,
allowing Python to be used in effect as a markup language. Also
supported are callbacks via hooks, recording and playback via
diversions, and dynamic, chainable filters. The system is highly
configurable via command line options and embedded commands.

http://www.alcyone.com/software/empy/

Note: the host variant is also included, as empy may be used by some
build systems (for example, the catkin build system used in Robot
Operating System).

Signed-off-by: Julien Olivain <ju.o@free.fr>