package/ros/angles: new package
This package provides a set of simple math utilities to work with
angles. The utilities cover simple things like normalizing an angle and
conversion between degrees and radians. But even if you're trying to
calculate things like the shortest angular distance between two joint
space positions of your robot, but the joint motion is constrained by
joint limits, this package is what you need. The code in this package is
stable and well tested.

https://wiki.ros.org/angles

Signed-off-by: Julien Olivain <ju.o@free.fr>