package/ros/nodelet_core: new ros package collection
Nodelet Core Metapackage: it contains necessary ros packages to work
with ROS nodelets. In ROS, each node is a process. A nodelet is instead
executed on a thread. The nodelet package is designed to provide a way
to run multiple algorithms in the same process with zero copy transport
between algorithms. This package provides both the nodelet base class
needed for implementing a nodelet, as well as the NodeletLoader class
used for instantiating nodelets.

https://wiki.ros.org/nodelet_core

Signed-off-by: Julien Olivain <ju.o@free.fr>