pkg-catkin: force PYTHON_EXECUTABLE to use host python3
On some build host, catkin package were detecting the target python
executable instead of the buildroot host python. This was leading to
build failures.

This commit forces the the python executable at the catkin package infra
level, to make sure catkin builds will always work.

Thanks to tiqo on #buildroot for reporting and helping on the issue.

Signed-off-by: Julien Olivain <ju.o@free.fr>